Update API version (#147)
* Bump VK Api version to 5.238
* Implemented new authorization flow (at the moment, without auto re-requesting token)
* Add support for sticker pack preview attachments
* Bump LongPoll to version 19
* Improved messages handling
* Fixed coloring issues
* Cache improvements
* Archive screen with full functionality
* Recomposition fixes
* Markdown support for messages bubbles
* Adjust app name font size based on screen width
* Navigation related improvements
* Add logout functionality

+++ core/database/src/main/kotlin/dev/meloda/fast/database/dao/MessageDao.kt
@@ -11,13 +11,13 @@ abstract class MessageDao : EntityDao<VkMessageEntity> {
     abstract suspend fun getAll(): List<VkMessageEntity>
 
     @Query("SELECT * FROM messages WHERE peerId IS (:conversationId)")
-    abstract suspend fun getAll(conversationId: Int): List<VkMessageEntity>
+    abstract suspend fun getAll(conversationId: Long): List<VkMessageEntity>
 
     @Query("SELECT * FROM messages WHERE id IN (:ids)")
     abstract suspend fun getAllByIds(ids: List<Int>): List<VkMessageEntity>
 
     @Query("SELECT * FROM messages WHERE id IS (:messageId)")
-    abstract suspend fun getById(messageId: Int): VkMessageEntity?
+    abstract suspend fun getById(messageId: Long): VkMessageEntity?
 
     @Query("DELETE FROM messages WHERE id IN (:ids)")
     abstract suspend fun deleteByIds(ids: List<Int>): Int
