+++ core/model/src/main/kotlin/dev/meloda/fast/model/api/data/VkArticleData.kt
@@ -6,7 +6,7 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class VkArticleData(
-    @Json(name = "id") val id: Int
+    @Json(name = "id") val id: Long
 ) : VkAttachmentData {
 
     fun toDomain(): VkArticleDomain = VkArticleDomain(
