+++ core/model/src/main/kotlin/dev/meloda/fast/model/api/data/VkWallReplyData.kt
@@ -5,12 +5,12 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class VkWallReplyData(
-    val id: Int,
-    val from_id: Int,
+    val id: Long,
+    val from_id: Long,
     val date: Int,
     val text: String,
-    val post_id: Int,
-    val owner_id: Int,
+    val post_id: Long,
+    val owner_id: Long,
     val parents_stack: List<Int>,
     val likes: Likes,
     val reply_to_user: Int?,
