+++ feature/messageshistory/src/main/kotlin/dev/meloda/fast/messageshistory/model/MessageDialog.kt
@@ -6,8 +6,8 @@ import dev.meloda.fast.model.api.domain.VkMessage
 @Immutable
 sealed class MessageDialog {
     data class MessageOptions(val message: VkMessage) : MessageDialog()
-    data class MessagePin(val messageId: Int) : MessageDialog()
-    data class MessageUnpin(val messageId: Int) : MessageDialog()
+    data class MessagePin(val messageId: Long) : MessageDialog()
+    data class MessageUnpin(val messageId: Long) : MessageDialog()
     data class MessageDelete(val message: VkMessage) : MessageDialog()
     data class MessagesDelete(val messages: List<VkMessage>) : MessageDialog()
